agent/pyproject.toml
[project]
name = "agent_voice_livekit_agent"
version = "0.0.1"
description = "An agent with streaming for Restack"
authors = [{ name = "Restack Team", email = "service@restack.io" }]
requires-python = ">=3.10,<3.14"
readme = "README.md"
dependencies = [
    "pydantic>=2.10.6",
    "watchfiles>=1.0.4",
    "python-dotenv==1.0.1",
    "openai>=1.61.0",
    "livekit-api>=0.8.2",
    "restack-ai==0.0.81",]

[project.scripts]
dev = "src.services:watch_services"
services = "src.services:run_services"
schedule = "schedule_agent:run_schedule_agent"
event = "event_agent:run_event_agent"

[tool.hatch.build.targets.sdist]
include = ["src"]

[tool.hatch.build.targets.wheel]
include = ["src"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
